requirements.txt
absl-py==0.8.1
altair==4.0.1
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
bleach==3.1.1
blinker==1.4
boto3==1.12.6
botocore==1.15.6
certifi==2019.11.28
channelnorm-cuda==0.0.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
correlation-cuda==0.0.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dlib==19.18.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
ffmpeg-python==0.1.17
Flask==1.1.1
Flask-Limiter==1.2.1
future==0.18.1
gast==0.2.2
google-pasta==0.1.8
grpcio==1.24.3
h5py==2.9.0
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.4.1
imgviz==0.12.1
importlib-metadata==1.5.0
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras==2.2.4
Keras-Applications==1.0.8
keras-contrib==2.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
labelme==4.2.9
limits==1.5.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
moviepy==1.0.1
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numpy==1.17.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
plaidml==0.6.0
plaidml-keras==0.5.0
proglog==0.1.9
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
pydeck==0.2.1
Pygments==2.4.2
pynvml==8.0.4
pyparsing==2.4.2
PyQt5==5.14.1
PyQt5-sip==12.7.1
pyrsistent==0.15.7
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.0
QtPy==1.9.0
requests==2.22.0
resample2d-cuda==0.0.0
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn==0.22.2
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
streamlit==0.56.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
tqdm==4.36.1
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.6
validators==0.14.2
watchdog==0.10.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.12.0
zipp==3.0.0
